type mismatch

+++ Top/csmodule.c
@@ -1310,9 +1310,8 @@ pvlock            pvlock.c
 vosim              vosim.c
 */
 
-typedef long long (*INITFN)(CSOUND *, void *);
+typedef long (*INITFN)(CSOUND *, void *);
 
-//extern const long (*babo_localops_init)(CSOUND *, OENTRY **);
 extern long babo_localops_init(CSOUND *, void *);
 extern long bilbar_localops_init(CSOUND *, void *);
 extern long compress_localops_init(CSOUND *, void *);
@@ -1322,9 +1321,10 @@ extern long eqfil_localops_init(CSOUND *, void *);
 extern int stdopc_ModuleInit(CSOUND *csound);
 extern int pvsopc_ModuleInit(CSOUND *csound);
 
-const INITFN staticmodules[] = {  babo_localops_init, bilbar_localops_init, vosim_localops_init,
-				  compress_localops_init, pvsbuffer_localops_init, eqfil_localops_init,
-                                 NULL };
+const INITFN staticmodules[] = {  babo_localops_init, bilbar_localops_init,
+                                  vosim_localops_init, compress_localops_init,
+                                  pvsbuffer_localops_init, eqfil_localops_init,
+                                  NULL };
 
 CS_NOINLINE int csoundInitStaticModules(CSOUND *csound)
 {
@@ -1344,10 +1344,10 @@ CS_NOINLINE int csoundInitStaticModules(CSOUND *csound)
       }
     }
     /* stdopc module */
-    if(stdopc_ModuleInit(csound)) return CSOUND_ERROR;
+    if (stdopc_ModuleInit(csound)) return CSOUND_ERROR;
 
     /* pvs module */
-    if(pvsopc_ModuleInit(csound)) return CSOUND_ERROR;
+    if (pvsopc_ModuleInit(csound)) return CSOUND_ERROR;
 
     /* module was initialised successfully */
     return CSOUND_SUCCESS;
